requirements.txt
python-telegram-bot
Pillow
opencv-python
numpy
